Refactor the activity queue
