+++ src/actix_web/inbox.rs
@@ -57,7 +57,7 @@ where
 mod test {
     use super::*;
     use crate::{
-        activity_queue::generate_request_headers,
+        activity_queue::request::generate_request_headers,
         config::FederationConfig,
         http_signatures::sign_request,
         traits::tests::{DbConnection, DbUser, Follow, DB_USER_KEYPAIR},
@@ -65,7 +65,6 @@ mod test {
     use actix_web::test::TestRequest;
     use reqwest::Client;
     use reqwest_middleware::ClientWithMiddleware;
-    use url::Url;
 
     #[tokio::test]
     async fn test_receive_activity() {
@@ -114,7 +113,7 @@ mod test {
 
     async fn setup_receive_test() -> (Bytes, TestRequest, FederationConfig<DbConnection>) {
         let inbox = "https://example.com/inbox";
-        let headers = generate_request_headers(&Url::parse(inbox).unwrap());
+        let headers = generate_request_headers(inbox).unwrap();
         let request_builder = ClientWithMiddleware::from(Client::default())
             .post(inbox)
             .headers(headers);
